# api.yaml
openapi: 3.0.0
info:
  title: ZenCura Hospital Management System API
  description: API documentation for the ZenCura HMS V2, allowing for patient management, appointment scheduling, and clinical reporting.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/reports/overview:
    get:
      summary: High-level management analytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Financial and trend data summaries